requirements.txt
pyyaml==6.0
schedule==1.1.0
requests==2.26.0
